move generic writables into its own package
