+++ schedoscope-export/src/main/java/org/schedoscope/export/writables/TextPairArrayWritable.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.schedoscope.export.ftp.outputformat;
+package org.schedoscope.export.writables;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -22,7 +22,6 @@ import java.util.List;
 import org.apache.hadoop.io.ArrayWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.Writable;
-import org.schedoscope.export.jdbc.outputformat.TextPairWritable;
 
 public class TextPairArrayWritable extends ArrayWritable {
